float horder

+++ poisson_batched_8_ORD_3D/CUDA/poisson_kernel_initialguess_cuda_kernel.cu
@@ -6,7 +6,7 @@ static int dims_poisson_kernel_initialguess_h[1][4] = {0};
 
 // user function
 
-__global__ void ops_poisson_kernel_initialguess(double *__restrict u_p,
+__global__ void ops_poisson_kernel_initialguess(float *__restrict u_p,
                                                 int bounds_0_l, int bounds_0_u,
                                                 int bounds_1_l, int bounds_1_u,
                                                 int bounds_2_l, int bounds_2_u,
@@ -23,15 +23,15 @@ __global__ void ops_poisson_kernel_initialguess(double *__restrict u_p,
 
   if (n_0 < bounds_0_u && n_1 < bounds_1_u && n_2 < bounds_2_u &&
       n_3 < bounds_3_u) {
-    ACC<double> u(dims_poisson_kernel_initialguess[0][0],
-                  dims_poisson_kernel_initialguess[0][1],
-                  dims_poisson_kernel_initialguess[0][2],
-                  u_p + n_0 + n_1 * dims_poisson_kernel_initialguess[0][0] +
-                      n_2 * dims_poisson_kernel_initialguess[0][0] *
-                          dims_poisson_kernel_initialguess[0][1] +
-                      n_3 * dims_poisson_kernel_initialguess[0][0] *
-                          dims_poisson_kernel_initialguess[0][1] *
-                          dims_poisson_kernel_initialguess[0][2]);
+    ACC<float> u(dims_poisson_kernel_initialguess[0][0],
+                 dims_poisson_kernel_initialguess[0][1],
+                 dims_poisson_kernel_initialguess[0][2],
+                 u_p + n_0 + n_1 * dims_poisson_kernel_initialguess[0][0] +
+                     n_2 * dims_poisson_kernel_initialguess[0][0] *
+                         dims_poisson_kernel_initialguess[0][1] +
+                     n_3 * dims_poisson_kernel_initialguess[0][0] *
+                         dims_poisson_kernel_initialguess[0][1] *
+                         dims_poisson_kernel_initialguess[0][2]);
 
     u(0, 0, 0) = 0.0;
   }
@@ -137,9 +137,9 @@ void ops_par_loop_poisson_kernel_initialguess_execute(
   }
 
   // set up initial pointers
-  double *__restrict__ u_p =
-      (double *)(args[0].data_d + args[0].dat->base_offset +
-                 blockidx_start * args[0].dat->batch_offset);
+  float *__restrict__ u_p =
+      (float *)(args[0].data_d + args[0].dat->base_offset +
+                blockidx_start * args[0].dat->batch_offset);
 
   int x_size = MAX(0, bounds_0_u - bounds_0_l);
   int y_size = MAX(0, bounds_1_u - bounds_1_l);
